refactor: group testing logic under own dir
