+++ src/lib/plugin/types/lens.ts
@@ -1,7 +1,7 @@
 import * as Logger from '@nexus/logger'
 import * as NexusSchema from '@nexus/schema'
 import * as Prompts from 'prompts'
-import * as Testing from '../../../runtime/testing'
+import * as Testing from '../../../testing/testing'
 import * as Layout from '../../layout'
 import * as PackageManager from '../../package-manager'
 import * as Process from '../../process'
