--- src/testing.ts
@@ -1,1 +1,0 @@
-export { createTestContext, TestContext } from './runtime/testing'
